requirements.txt
fastapi
uvicorn
python-dotenv
pyyaml
langchain
langchain-community
langchain-core
langchain-openai
langchain_google_genai
langchain-anthropic
langchain-chroma
chromadb
langchain-huggingface
guardrails-ai
docx2txt
unstructured
markdown
redis